Split up responses
