+++ fs/directory_listing.go
@@ -4,8 +4,6 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"mime"
-	"os"
 	"path"
 	"strconv"
 
@@ -67,33 +65,4 @@ func makeListItem(text string) string {
 func (listing DirectoryListing) makeLink(filename string) string {
 	href := path.Join(listing.HrefPrefix, filename)
 	return fmt.Sprintf("<a href=\"%s\">%s</a>", href, filename)
-}
-
-type FileContents struct {
-	Filename string
-}
-
-func (contents FileContents) WriteTo(client io.Writer) error {
-	response.WriteStatusLine(client, 200, "OK")
-	contents.writeHeadersDescribingFile(client)
-	response.WriteEndOfMessageHeader(client)
-
-	file, _ := os.Open(contents.Filename)
-	response.CopyToBody(client, file)
-	return nil
-}
-
-func (contents FileContents) writeHeadersDescribingFile(client io.Writer) {
-	response.WriteHeader(client, "Content-Type", contentTypeFromFileExtension(contents.Filename))
-	info, _ := os.Stat(contents.Filename)
-	response.WriteHeader(client, "Content-Length", strconv.FormatInt(info.Size(), 10))
-}
-
-func contentTypeFromFileExtension(filename string) string {
-	extension := path.Ext(filename)
-	if extension == "" {
-		return "text/plain"
-	}
-
-	return mime.TypeByExtension(extension)
 }
